moved settings to their own file - added reflection for reading and writing settings, so adding a new one property to the object will automatically result in it being included in files. - need to figure out how to reflect into a form so that all settings (perhaps except version) are listed automatically?
